transfered into the correct files.  ChangeType is solved.
