+++ src/components/TwoDice.tsx
@@ -11,6 +11,26 @@ export function d6(): number {
     return 1 + Math.floor(Math.random() * 6);
 }
 
+const [left_die, setLeftDiceValue] = useState<number>(0);
+const [right_die, setRightDiceValue] = useState<number>(0);
+
+function rollLeftDice() {
+    setLeftDiceValue(d6());
+    left_die;
+}
+function rollRightDice() {
+    setRightDiceValue(d6());
+    right_die;
+}
+
 export function TwoDice(): JSX.Element {
-    return <div>Two Dice</div>;
+    return (
+        <div>
+            <div>Two Dice</div>;
+            <Button onClick={rollLeftDice}>Roll Left</Button>;
+            {<span>{left_die}</span>}
+            <Button onClick={rollRightDice}>Roll Right</Button>;
+            {<span>{right_die}</span>}
+        </div>
+    );
 }
